requirements.txt
# requirements.txt
# Pinned dependencies for CI stability - compatible versions

fastapi==0.117.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests==2.32.4
aiohttp==3.12.14
psycopg2-binary==2.9.9
redis==4.6.0
celery==5.3.4
prometheus-client==0.19.0
aiokafka==0.8.1  # Optional: for Kafka streaming support

# Security dependencies (v1.2.1+)
python-jose[cryptography]==3.3.0  # JWT token handling
passlib[bcrypt]==1.7.4            # Password hashing
bcrypt==4.1.2                      # Password hashing backend
cryptography==41.0.7               # Cryptographic operations

# AI/ML dependencies (v1.2.0+)
# Note: These are optional - DECOYABLE works without them using pattern-based analysis
# transformers>=4.36.0  # Uncomment for Phi-3 local AI support
# torch>=2.1.0          # Uncomment for Phi-3 local AI support
# ollama>=0.1.0         # Uncomment for Ollama client (or use direct HTTP)
numpy>=1.24.0          # Required for AI predictions
